move types tests to tests directory
